% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_aquire.R
\name{sco_acquire}
\alias{sco_acquire}
\title{acquire data and rearrange columns}
\usage{
sco_acquire(year_input, league_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a data frame
}
\description{
acquire data and rearrange columns
}
