% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_bayesian_btts.R
\name{sco_bayesian_btts}
\alias{sco_bayesian_btts}
\title{bayesian model of btts based on logit prior distirubtion}
\usage{
sco_bayesian_btts(year_input, league_input, var_input, team_input, plot_content,
  tester)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{var_input}{select whether team is play home ("hometeam") or away ("awayteam")}

\item{team_input}{team}

\item{plot_content}{TRUE/FALSE to plot data}

\item{tester}{TRUE/FALSE whether data is used for testing (exclude last 30 days results)}
}
\value{
a tibble of team and median of samples distributions ("best guess")
}
\description{
bayesian model of btts based on logit prior distirubtion
}
