% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_bayesian_btts_calibration.R
\name{sco_bayesian_calibration}
\alias{sco_bayesian_calibration}
\title{calibrate the bayesian model by changing the threshold values}
\usage{
sco_bayesian_calibration(league_input, year_input)
}
\arguments{
\item{league_input}{co_pl (Premiership) or sco_ch (Championship)}

\item{year_input}{input as character in years of seasons, eg. "yyyy"}
}
\value{
a object of prediction strength for combination of upper nad lower treshold values
}
\description{
calibrate the bayesian model by changing the threshold values
}
