% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_bayesian_btts_tester.R
\name{sco_bayesian_tester}
\alias{sco_bayesian_tester}
\title{calibrate the bayesian model with respect to where to put best guess thresh}
\usage{
sco_bayesian_tester(league_input, year_input, play_upper_treshhold,
  play_lower_treshhold)
}
\arguments{
\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{play_upper_treshhold}{thresh of best guess to accept play of case}

\item{play_lower_treshhold}{thresh of best guess to accept play of non-case}
}
\value{
an object
}
\description{
calibrate the bayesian model with respect to where to put best guess thresh
}
