% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_bet_acquire.R
\name{sco_bet_acquire}
\alias{sco_bet_acquire}
\title{Acquire betting data from Bet365}
\usage{
sco_bet_acquire(year_input, league_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a data frame
}
\description{
Acquire betting data from Bet365
}
