% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_bet_plot.R
\name{sco_bet_ou_plot}
\alias{sco_bet_ou_plot}
\title{Analyse input of odds: how is the input odds compared to data from this season?}
\usage{
sco_bet_ou_plot(year_input, league_input, team_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{team_input}{selected_team}
}
\value{
a ggplot2 object
}
\description{
Analyse input of odds: how is the input odds compared to data from this season?
}
