% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_bet_plot.R
\name{sco_bet_plot}
\alias{sco_bet_plot}
\title{Analyse input of odds: how is the input odds compared to data from this season?}
\usage{
sco_bet_plot(venue, venue_naming, team, odds)
}
\arguments{
\item{venue}{hometeam or awayteam not in ""}

\item{venue_naming}{hometeam or awayteam as character (in "")}

\item{team}{team to look at (in "")}

\item{odds}{current odds to compare with}

\item{year_input}{input as character in years of seasons, eg. "yyyy"}
}
\value{
a ggplot2 object
}
\description{
Analyse input of odds: how is the input odds compared to data from this season?
}
