% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_corner_plot.R
\name{sco_corner_plot}
\alias{sco_corner_plot}
\title{Corner plot}
\usage{
sco_corner_plot(team_input, line_on)
}
\arguments{
\item{team_input}{team to look at}

\item{line_on}{line to compare with}
}
\value{
a ggplot2 object
}
\description{
Corner plot
}
