% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_stats_per_minute.R
\name{sco_goal_per_minute}
\alias{sco_goal_per_minute}
\title{Goals per minute}
\usage{
sco_goal_per_minute(year_input, league_input, min_goals, min_matches)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{min_goals}{minimum goals to be scored to be shown}

\item{min_matches}{minimum matches played to be shown}
}
\value{
a dataframe of player data


}
\description{
Goals per minute
}
