% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_match_cards.R
\name{sco_match_cards}
\alias{sco_match_cards}
\title{acquire match card statistics}
\usage{
sco_match_cards(
  year_input,
  league_input,
  hometeam_input,
  awayteam_input,
  referee_input,
  card_line_input
)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{hometeam_input}{hometeam}

\item{referee_input}{referee name (D Robertson)}

\item{card_line_input}{card line input}

\item{awayteam_imput}{awayteam}
}
\value{
a data frame
}
\description{
acquire match card statistics
}
