% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_misc_stats_develop.R
\name{sco_misc_stats_develop}
\alias{sco_misc_stats_develop}
\title{Misc stats (goals,shots, cards, corners) development during year}
\usage{
sco_misc_stats_develop(stat_input, year_input, league_input)
}
\arguments{
\item{stat_input}{stat devleopment (shots, shontsontarget,corners,yellowcard,redcards)}

\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a ggplot2 object of development of \code{stat_input_home} and \code{stat_input_away} during \code{year_input}
}
\description{
Misc stats (goals,shots, cards, corners) development during year
}
