% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_player_contribution_matrix.R
\name{sco_player_contribution_matrix}
\alias{sco_player_contribution_matrix}
\title{Goal Contribution Matrix}
\usage{
sco_player_contribution_matrix(year_input, league_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a dataframe of player data
}
\description{
Goal Contribution Matrix
}
