% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_player_data.R
\name{sco_player_data}
\alias{sco_player_data}
\title{Player data from scottish divisions}
\usage{
sco_player_data(year_input, league_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a dataframe of player data
}
\description{
Player data from scottish divisions
}
