% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_poisson_xg.R
\name{sco_poisson_xg}
\alias{sco_poisson_xg}
\title{possion model of expected goals for home and away team}
\usage{
sco_poisson_xg(relative_strength_input)
}
\arguments{
\item{relative_strength_input}{input object from \code{sco_relative_strength}}
}
\value{
a object with prediction
}
\description{
possion model of expected goals for home and away team
}
