% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_poisson_xg_calibration.R
\name{sco_poisson_xg_calibration}
\alias{sco_poisson_xg_calibration}
\title{test actual data against model data for difference previous games}
\usage{
sco_poisson_xg_calibration(
  year_input,
  league_input,
  n_previous_game_low,
  n_previous_game_high
)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{n_previous_game_low}{number of games to base the relative strength upon (lower boundary)}

\item{n_previous_game_high}{number of games to base the relative strength upon (upper boundary)
#'}
}
\value{
a plot that shows the highest rate of trues (result hits) in terms of looking games back
}
\description{
test actual data against model data for difference previous games
}
