% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_poisson_xg_predictions.R
\name{sco_poisson_xg_predictions}
\alias{sco_poisson_xg_predictions}
\title{plot poisson model}
\usage{
sco_poisson_xg_predictions(relative_strength_input)
}
\arguments{
\item{relative_strength_input}{input object from \code{sco_relative_strength}}
}
\value{
a data frame object to plot result of model
}
\description{
plot poisson model
}
