% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_poisson_xg_tester.R
\name{sco_poisson_xg_tester}
\alias{sco_poisson_xg_tester}
\title{test actual data against model data}
\usage{
sco_poisson_xg_tester(year_input, league_input, n_previous_games)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{n_previous_games}{number of games to base the relative strength upon}
}
\value{
an object that compares played games versus predictions for one value of \code{n_previous_games}
}
\description{
test actual data against model data
}
