% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_ref_card_plot.R
\name{sco_ref_card_plot}
\alias{sco_ref_card_plot}
\title{Plot cards given by each referee in a season}
\usage{
sco_ref_card_plot(year_input, league_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a ggplot2 object of cards by refs in \code{year_input} of \code{league_input}
}
\description{
Plot cards given by each referee in a season
}
