% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_standing_simulator.R
\name{sco_standings}
\alias{sco_standings}
\title{standings t at the given time}
\usage{
sco_standings(year_input, league_input, at_time_date)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{at_time_date}{time at standing "yyyy-mm-dd" format}
}
\value{
table standing at the \code{at_time_date}
}
\description{
standings t at the given time
}
