% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_team_cards.R
\name{sco_team_cards}
\alias{sco_team_cards}
\title{acquire match card statistics}
\usage{
sco_team_cards(year_input, league_input, team_input, card_line_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{team_input}{hometeam}

\item{card_line_input}{card line input}
}
\value{
a data frame
}
\description{
acquire match card statistics
}
