% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_team_corners.R
\name{sco_team_corners}
\alias{sco_team_corners}
\title{acquire team corner statistics}
\usage{
sco_team_corners(year_input, league_input, team_input, corner_line_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{team_input}{hometeam}

\item{card_line_input}{card line input}
}
\value{
a data frame
}
\description{
acquire team corner statistics
}
