% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_team_stats.R
\name{sco_team_stats}
\alias{sco_team_stats}
\title{Title}
\usage{
sco_team_stats(year_input, league_input)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a stat sheet
}
\description{
Title
}
