% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_bet_acquire.R
\name{sco_bet_acquire}
\alias{sco_bet_acquire}
\title{Acquire betting data from Bet365}
\usage{
sco_bet_acquire(league_input)
}
\arguments{
\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}
}
\value{
a data frame
}
\description{
Acquire betting data from Bet365
}
