% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sco_poisson_xg.R
\name{sco_poisson_xg}
\alias{sco_poisson_xg}
\title{possion model of expected goals for home and away team}
\usage{
sco_poisson_xg(
  year_input,
  league_input,
  hometeam_input,
  awayteam_input,
  n_previous_games,
  date_lim
)
}
\arguments{
\item{year_input}{input as character in years of seasons, eg. "yyyy"}

\item{league_input}{sco_pl (Premiership) or sco_ch (Championship)}

\item{hometeam_input}{hometeam}

\item{awayteam_input}{awayteam}

\item{n_previous_games}{number of games to base the relative strength upon}

\item{date_lim}{limit of date to include}
}
\value{
a object with prediction
}
\description{
possion model of expected goals for home and away team
}
